LayoutTests/http/tests/w3c/html/semantics/embedded-content/the-area-element/area-stringifier-expected.txt


PASS HTMLAreaElement stringifier 
FAIL HTMLAreaElement stringifier 1 assert_throws: function "function () {
        aObject.toString.call(v);
      }" did not throw
FAIL HTMLAreaElement stringifier 2 assert_throws: function "function () {
        aObject.toString.call(v)
      }" did not throw
PASS HTMLAreaElement stringifier 3 
FAIL HTMLAreaElement stringifier 4 assert_equals: expected "" but got "[object HTMLAreaElement]"
FAIL HTMLAreaElement stringifier 5 assert_equals: expected "" but got "[object HTMLAreaElement]"
FAIL HTMLAreaElement stringifier 6 assert_throws: function "function () {
        aObject.toString.call(v);
      }" did not throw
FAIL HTMLAreaElement stringifier 7 assert_throws: function "function () {
        aObject.toString.call(v)
      }" did not throw
PASS HTMLAreaElement stringifier 8 
FAIL HTMLAreaElement stringifier 9 assert_equals: expected "http://127.0.0.1:8000/w3c/html/semantics/embedded-content/the-area-element/foo" but got "[object HTMLAreaElement]"
FAIL HTMLAreaElement stringifier 10 assert_equals: expected "http://127.0.0.1:8000/w3c/html/semantics/embedded-content/the-area-element/foo" but got "[object HTMLAreaElement]"
